gRPC connection with node.js-MongoDB-caching-server & gRPC communication implemented [ 'proposal_event_message' ]

+++ src/main/java/org/zerock/voteservice/dto/event/ExpiredPendingEventDto.java
@@ -14,7 +14,7 @@ public class ExpiredPendingEventDto {
     private int voteCount;
 
     @JsonProperty("vote_option_counts")
-    private Map<String, Long> voteOptionCounts;
+    private Map<String, Integer> voteOptionCounts;
 }
